+++ app/src/main/java/info/gdbtech/greg/myapplicationii/ui/main/Collision.kt
@@ -3,18 +3,18 @@ package info.gdbtech.greg.myapplicationii.ui.main
 import android.graphics.Canvas
 import android.graphics.ColorFilter
 
-class Collision(pointMassA: PointMass, pointMassB: PointMass, var shortLimit: Double) :
+class Collision(pointMassA: PointMass, pointMassB: PointMass, var shortLimit: Float) :
     Force(pointMassA, pointMassB) {
 
     private var delta = pointMassB.pos - pointMassA.pos
 
-    var longLimit: Double = Double.POSITIVE_INFINITY
+    var longLimit: Float = Float.POSITIVE_INFINITY
 
     private var slSquared = shortLimit * shortLimit
 
     private var llSquared = longLimit * longLimit
 
-    override fun scale(scaleFactor: Double) {
+    override fun scale(scaleFactor: Float) {
         shortLimit *= scaleFactor
         longLimit *= scaleFactor
     }
@@ -22,13 +22,13 @@ class Collision(pointMassA: PointMass, pointMassB: PointMass, var shortLimit: Do
     override fun sc(env: Environment) {
         val dp = delta.dotProd(delta)
         if (dp < slSquared) {
-            val scaleFactor = slSquared / (dp + slSquared) - 0.5
+            val scaleFactor = slSquared / (dp + slSquared) - 0.5f
             delta.scale(scaleFactor)
             pointMassA.pos.sub(delta)
             pointMassB.pos.sub(delta)
         }
         if (dp > llSquared) {
-            val scaleFactor = llSquared / (dp + llSquared) - 0.5
+            val scaleFactor = llSquared / (dp + llSquared) - 0.5f
             delta.scale(scaleFactor)
             pointMassA.pos.sub(delta)
             pointMassB.pos.sub(delta)
